+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 
   * replaced `stubbles\db\config\DatabaseConfigReader` with `stubbles\db\config\DatabaseConfigurations`
   * replaced `stubbles\db\ioc\ConnectionProvider` with `stubbles\db\\DatabaseConnections`
+  * replaced `stubbles\db\ioc\DatabaseProvider` with `stubbles\db\\Databases`
   * removed `stubbles\db\ioc\DatabaseBindingModule`, did not provide any additional value
 
 ### Other changes